ApiFastAPI/pyproject.toml
[project]
name = "apifastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel Guerra",email = "dev.gguerra@gmail.com"}
]
license = {text = "APACHE 2.0"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "dependency-injector (>=4.48.1,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "pytest-coverage (>=0.0,<0.1)"
]


[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
